+++ src/market_sentinel_ai/alerts/__init__.py
@@ -1,0 +1,4 @@
+from market_sentinel_ai.alerts.channels import DryRunAlertChannel, JsonlAlertChannel
+
+__all__ = ["DryRunAlertChannel", "JsonlAlertChannel"]
+